requirements.txt
Adafruit-Blinka==3.5.2
adafruit-circuitpython-bme280==2.3.3
adafruit-circuitpython-busdevice==4.1.1
adafruit-circuitpython-hcsr04==0.3.6
Adafruit-PlatformDetect==1.4.5
Adafruit-PureIO==1.0.4
future==0.18.2
iso8601==0.1.12
pkg-resources==0.0.0
pyftdi==0.42.2
pyserial==3.4
pyusb==1.0.2
PyYAML==5.3
rpi-ws281x==4.2.3
RPi.GPIO==0.7.0
serial==0.0.97
spidev==3.4
sysv-ipc==1.0.1
